fix: surface shortcut registration failures
